requirements.txt
# Alembic: migrații de schemă și versiuni. Rol: gestionare migrații DB. Alternative: Django migrations, Flyway.
alembic==1.16.1

# annotated-types: tipuri adiționale pentru validări. Rol: tipuri pentru Pydantic. Alternative: tipuri built-in.
annotated-types==0.7.0

# anyio: abstracție I/O asincron. Rol: suport asinc pentru FastAPI. Alternative: Trio, curio.
anyio==4.9.0

# bcrypt: hashing parole. Rol: securizarea parolelor. Alternative: Argon2, PBKDF2.
bcrypt==4.3.0

# boto3 & botocore: SDK AWS. Rol: acces la servicii AWS (S3, ECR, RDS). Alternative: AWS CLI.
boto3==1.37.23
botocore==1.37.23

# cffi: binding C pentru criptografie. Rol: suport pentru cryptography, bcrypt. Alternative: pycryptodome.
cffi==1.17.0

# click: CLI builder. Rol: argumente/subcomenzi scripturi. Alternative: argparse, Typer.
click==8.1.8

# cryptography: criptografie avansată. Rol: SSL, criptare simetrică/asimetrică. Alternative: pycryptodome.
cryptography==39.0.1

# dnspython: rezolvare DNS. Rol: validări email și interacțiuni DNS. 
dnspython==2.7.0

# ecdsa: semnături digitale. Rol: ECDSA pentru criptografie. Alternative: rsa.
ecdsa==0.19.1

# email_validator: validare email. Rol: asigură format corect al adreselor. Alternative: regex.
email_validator==2.2.0

# fastapi: framework web asincron. Rol: API HTTP și WebSocket. Alternative: Flask, Django.
fastapi==0.115.12

# h11, h2, hpack: protocoale HTTP/1.1 și HTTP/2. Rol: gestionare conexiuni HTTP. 
h11==0.14.0
h2==4.2.0
hpack==4.1.0

# httptools: parsing HTTP performant. Rol: accelerează parsing la Uvicorn. 
httptools==0.6.4

# Hypercorn & hyperframe: server ASGI cu HTTP/2. Rol: server de producție alternativ la Uvicorn.
Hypercorn==0.17.3
hyperframe==6.1.0

# idna & sniffio: suport DNS și handshake-uri. Rol: compatibilitate rețea.
idna==3.10
sniffio==1.3.1

# Mako & MarkupSafe: template engine. Rol: generare șabloane e-mail sau HTML. 
Mako==1.3.10
MarkupSafe==3.0.2

# passlib: hashing parole cu interfață. Rol: wrapper pentru algoritmi de hashing. 
passlib==1.7.4

# priority: pachet specific de priorități în taskuri.
priority==2.0.0

# pyasn1 & pycparser: ASN.1 și analizator C. Rol: suport criptare și binding C.
pyasn1==0.4.8
pycparser==2.22

# pydantic & pydantic_core: modelare și validare date. Rol: schema și validare. Alternative: Marshmallow.
pydantic==2.11.1
pydantic_core==2.33.0

# PyMySQL: driver MySQL pur Python. Rol: conexiune DB. Alternative: mysqlclient (C).
PyMySQL==1.1.1

# python-dateutil: manipulare date și intervale. Rol: parsing și formatare date.
python-dateutil==2.9.0.post0

# python-dotenv: încărcare variabile mediu. Rol: citire .env pentru config.
python-dotenv==1.1.0

# python-jose: JWT. Rol: generare/validare token-uri. Alternative: PyJWT.
python-jose==3.4.0

# PyYAML: parsing YAML. Rol: citire config YAML.
PyYAML==6.0.2

# rsa: criptografie RSA. Rol: implementare RSA. Alternative: cryptography.
rsa==4.9.1

# s3transfer: transfer optimizat S3. Rol: multipart și retry.
s3transfer==0.11.4

# six, typing-inspection, typing_extensions: suport typing. Rol: compatibilitate Python typing.
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.0

# urllib3: client HTTP robust. Rol: HTTP requests. Alternative: requests.
urllib3==2.3.0

# uvicorn, uvloop: server ASGI și loop asincron performant. Rol: dezvoltare și producție rapidă.
uvicorn==0.34.0
uvloop==0.21.0

# watchfiles: reload dezvoltare. Rol: live reload la schimbare fișiere.
watchfiles==1.0.5

# websockets & wsproto: comunicații WebSocket. Rol: real-time messaging. Alternative: Socket.IO.
websockets==15.0.1
wsproto==1.2.0